+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## 08-09-2017
+### Added
+- functional: either.
+
 ## 07-09-2017
 ### Added
 - Basics: classes and enumconstants.